requirements.txt
langgraph==0.6.6
langchain-openai
python-dotenv
beautifulsoup4
lxml